WIP: Move QueriesExt out of apps
